:recycle:  split graphql module in submodules
