+++ packages/utils/src/graphql/group.ts
@@ -14,12 +14,12 @@ import type {
   Maybe,
 } from "@graphql-markdown/types";
 
-import { hasAstNode } from "./graphql";
+import { hasAstNode } from "./introspection";
 
 /**
  * Parses a GraphQL schema to build a map of entities with matching `groupByDirective` option.
  *
- * @param schemaMap - the GraphQL schema map returned by {@link graphql.getSchemaMap}
+ * @param schemaMap - the GraphQL schema map returned by {@link getSchemaMap}
  * @param groupByDirective - the `groupByDirective` option.
  *
  * @returns a map of entities with matching group name.
