+++ packages/utils/src/graphql/index.ts
@@ -1,5 +1,7 @@
-export * from "./graphql";
 export * from "./directive";
 export * from "./formatter";
 export * from "./group";
+export * from "./guard";
+export * from "./introspection";
+export * from "./loader";
 export * from "./relation";
