+++ packages/utils/src/index.ts
@@ -4,8 +4,5 @@ export * from "./array";
 export * from "./object";
 export * from "./string";
 export * as pathUrl from "./url";
-export * from "./graphql/graphql";
-export * from "./graphql/group";
-export * from "./graphql/directive";
+export * from "./graphql";
 export * from "./logger";
-export * from "./graphql/formatter";
